requirements.txt
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.2
PyMySQL==0.7.9
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.11
